pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "docu-flow"
version = "0.1.0"
description = "Automated clinical trial eligibility screening via protocol PDF parsing"
requires-python = ">=3.11"
dependencies = [
    # PDF extraction — pymupdf4llm wraps pymupdf and outputs LLM-optimized Markdown.
    # [ocr,layout] enables auto-OCR fallback (Tesseract) + multi-column reading-order fix.
    # pdfplumber kept as fallback for complex table structures (merged cells).
    "pymupdf>=1.24.0",
    "pymupdf4llm[ocr,layout]>=0.3.4",
    "pdfplumber>=0.11.0",
    "pytesseract>=0.3.13",
    "Pillow>=10.3.0",
    "pdf2image>=1.17.0",
    # Text repair — run ftfy immediately after extraction to fix mojibake,
    # garbled encodings, and PDF extraction artifacts before any NLP or LLM call.
    "ftfy>=6.3.0",
    # Fuzzy matching — rapidfuzz replaces naive str.lower() header matching.
    # Handles OCR typos in section headers ("Eligibllity Criterla" → match).
    "rapidfuzz>=3.14.0",
    # LLM — 0.84.0 adds: messages.parse() for native structured output via Pydantic,
    # count_tokens() for pre-flight budget checks, and native PDF content blocks.
    # tiktoken removed: Anthropic SDK now provides native token counting.
    "anthropic>=0.84.0",
    "openai>=1.45.0",           # optional fallback
    # API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",  # file uploads
    # Data / validation
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    # Storage / queue
    "redis>=5.0.8",
    "celery[redis]>=5.4.0",
    # Utilities
    "structlog>=24.4.0",
    "tenacity>=9.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "rich>=13.8.0",
]

[project.optional-dependencies]
# lingua: non-English protocol detection. 96 MB wheel — only install if
# processing multi-language trial protocols.
lingua = [
    "lingua-language-detector>=2.1.0",
]
# pdf-strategies: extra deps for the parallel strategy test harness (S4-Unstructured)
pdf-strategies = [
    "unstructured[pdf]>=0.15.0",
]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",            # async test client
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "pre-commit>=3.8.0",
    "google-generativeai>=0.8.0",  # Gemini API for integration tests
]

[project.scripts]
docu-flow = "docu_flow.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/docu_flow"]

# ---------------------------------------------------------------------------
# Ruff (linter + formatter)
# ---------------------------------------------------------------------------
[tool.ruff]
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["docu_flow"]

# ---------------------------------------------------------------------------
# Mypy
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-q"
markers = [
    "integration: marks tests as integration tests that call live APIs (deselect with '-m not integration')",
    "slow: marks tests as slow (large file processing, extended OCR); deselect with '-m not slow'",
    "anthropic: requires ANTHROPIC_API_KEY to be set",
    "google: requires GOOGLE_API_KEY to be set",
]
